Minor changes for clean-ish check
The imported SMC functions have been cleaned to achieve a devtools::check() without errors or warnings. Several functions in the post_processing_functions.R file have been commented out and shall be reincorporated at a later stage. The documentation for those functions might be worked on a later commit to allow the test scripts to work.