Adjusting test expectations to new code output
The differing output is caused by `set.seed(101)` having a different effect on the new C++ functions, particularly the output of `test_1_forward$aug_ranking`, which trickles down into the calculation of `test_1_backward_b`, thus causing different probabilities. The code was back-checked with the original R code, in which case the output seems the same. However, this should be checked with Anja.