Opening CI to all branches
r.yml is currently configured for only triggering GitHub Actions CI
workflow for the `master` and the `main` (inexistent here) branches.
Opening it up to any branch so commits to `smc-mallows` also get
triggered. This is important so we can better foresee problems in the
upcoming merge with `master`. Opening up to any branch name shouldn't be
a problem on this repository, since `master` is the only long-lived
branch at the moment.