Removed gtools dependency on SMC (#142)
Removed dependency on gtools::mixedorder

Works by extracting everything that isn't a digit ([0-9]) and ultimately
using the index of the sorted elements to sort `rho_posterior_interval`.
Works well for the standard "Item DD" format that SMC uses, and probably
for other formats as well, but this wasn't extensively tested.