Fixed calculation of ESS_vec elements
`^` is not a valid Armadillo operator (see list
[here](http://arma.sourceforge.net/docs.html#operators)). The usual way
to square a matrix is to multiply it by its transpose, so that's what I
proposed here. The operation was also separated from the sum so arma would stop
complaining (maybe this could be a one-liner again by using `conv_to()`,
but I didn't wanna do that.