Update smc_mallows_new_users_complete.cpp (#189)
* Update smc_mallows_new_users_complete.cpp

Include a vector named ESS_vec (a vector of length Time) which contains the Effective Sample Size of the particles in each iteration.

* Update smc_mallows_new_users_complete.cpp

* Update smc_mallows_new_users_complete.cpp

* Update smc_mallows_new_users_complete.cpp

* Fixed calculation of ESS_vec elements

`^` is not a valid Armadillo operator (see list
[here](http://arma.sourceforge.net/docs.html#operators)). The usual way
to square a matrix is to multiply it by its transpose, so that's what I
proposed here. The operation was also separated from the sum so arma would stop
complaining (maybe this could be a one-liner again by using `conv_to()`,
but I didn't wanna do that.

* Fixup for 06d7c87

Previous commit was syntatically OK, but issues arised when running tests. This fixes it

* Adjusted unit tests to ESS addition

* Update smc_mallows_new_users_complete.cpp

Changed the calculation for ESS - I was using the incorrect formula.

* Added unit test for ESS

Co-authored-by: Waldir Leoncio <w.l.netto@medisin.uio.no>